Parser Update and Field.java added
Added opportunity to evaluate expression with 3 variables of type
‘Double’
Class responsible for electric field added.

+++ src/ru/ifmo/ctddev/bandarchuk/expression/BinaryOperation.java
@@ -23,7 +23,7 @@ public abstract class BinaryOperation implements Operand {
     }
 
     @Override
-    public int evaluate(int x, int y, int z) throws OverflowException, DivisionByZeroException {
+    public double evaluate(double x, double y, double z) throws OverflowException, DivisionByZeroException {
         return operation(firstOperand.evaluate(x, y, z), secondOperand.evaluate(x, y, z));
     }
