+++ src/ru/ifmo/ctddev/bandarchuk/expression/Const.java
@@ -22,7 +22,7 @@ public class Const implements Operand {
     }
 
     @Override
-    public int evaluate(int x, int y, int z) {
-        return (int)value;
+    public double evaluate(double x, double y, double z) {
+        return value;
     }
 }
