+++ src/ru/ifmo/ctddev/bandarchuk/expression/TripleExpression.java
@@ -4,5 +4,5 @@ package ru.ifmo.ctddev.bandarchuk.expression;
  * Created by YuryBandarchuk on 3/22/17.
  */
 public interface TripleExpression {
-    int evaluate(int x, int y, int z) throws OverflowException, DivisionByZeroException;
+    double evaluate(double x, double y, double z) throws OverflowException, DivisionByZeroException;
 }
